delete file only if backing up to a cloud storage

+++ core/download_thread.py
@@ -69,7 +69,7 @@ class DownloadThread(Thread):
                 self.storage.upload_file(filename, path)
             is_successful = True
 
-        if filename:
+        if self.storage and filename:
             os.remove(path)
 
         return is_successful
